+++ mayan/apps/metadata/literals.py
@@ -1,0 +1,5 @@
+from .parsers import MetadataParser
+from .validators import MetadataValidator
+
+DEFAULT_METADATA_AVAILABLE_VALIDATORS = MetadataValidator.get_import_paths()
+DEFAULT_METADATA_AVAILABLE_PARSERS = MetadataParser.get_import_paths()