+++ mayan/apps/storage/literals.py
@@ -1,4 +1,14 @@
+import os
+import tempfile
+
+from django.conf import settings
+
 DEFAULT_STORAGE_BACKEND = 'django.core.files.storage.FileSystemStorage'
+DEFAULT_STORAGE_SHARED_STORAGE = 'django.core.files.storage.FileSystemStorage'
+DEFAULT_STORAGE_SHARED_STORAGE_ARGUMENTS = {
+    'location': os.path.join(settings.MEDIA_ROOT, 'shared_files')
+}
+DEFAULT_STORAGE_TEMPORARY_DIRECTORY = tempfile.gettempdir()
 DELETE_STALE_UPLOADS_INTERVAL = 60 * 10  # 10 minutes
 MSG_MIME_TYPES = (
     'application/vnd.ms-outlook', 'application/vnd.ms-office'