+++ mayan/apps/task_manager/settings.py
@@ -10,31 +10,31 @@ from .literals import (
 namespace = SettingNamespace(label=_('Celery'), name='celery')
 
 setting_celery_broker_login_method = namespace.add_setting(
-    global_name='CELERY_BROKER_LOGIN_METHOD',
-    default=DEFAULT_CELERY_BROKER_LOGIN_METHOD, help_text=_(
+    default=DEFAULT_CELERY_BROKER_LOGIN_METHOD,
+    global_name='CELERY_BROKER_LOGIN_METHOD', help_text=_(
         'Default: "AMQPLAIN". Set custom amqp login method.'
-    ),
+    )
 )
 setting_celery_broker_url = namespace.add_setting(
-    global_name='CELERY_BROKER_URL', default=DEFAULT_CELERY_BROKER_URL,
+    default=DEFAULT_CELERY_BROKER_URL, global_name='CELERY_BROKER_URL',
     help_text=_(
         'Default: "amqp://". Default broker URL. This must be a URL in '
         'the form of: transport://userid:password@hostname:port/virtual_host '
         'Only the scheme part (transport://) is required, the rest is '
         'optional, and defaults to the specific transports default values.'
-    ),
+    )
 )
 setting_celery_broker_use_ssl = namespace.add_setting(
-    global_name='CELERY_BROKER_USE_SSL',
-    default=DEFAULT_CELERY_BROKER_USE_SSL, help_text=_(
+    default=DEFAULT_CELERY_BROKER_USE_SSL,
+    global_name='CELERY_BROKER_USE_SSL', help_text=_(
         'Default: "Disabled". Toggles SSL usage on broker connection '
         'and SSL settings. The valid values for this option vary by '
         'transport.'
-    ),
+    )
 )
 setting_celery_result_backend = namespace.add_setting(
-    global_name='CELERY_RESULT_BACKEND',
-    default=DEFAULT_CELERY_RESULT_BACKEND, help_text=_(
+    default=DEFAULT_CELERY_RESULT_BACKEND,
+    global_name='CELERY_RESULT_BACKEND', help_text=_(
         'Default: No result backend enabled by default. The backend used '
         'to store task results (tombstones). Refer to '
         'http://docs.celeryproject.org/en/v4.1.0/userguide/configuration.'